node-postgres: Overhaul and fix CSV import test case
The test was completely defunct before as it didn't honour the
asynchronous runtime behaviour correctly. Now, it has been
rewritten to use async/await and Promises appropriately.
