+++ tests/client_tests/node-postgres/package.json
@@ -7,7 +7,7 @@
     "url": "https://github.com/crate/crate-qa.git",
     "directory": "tests/client_tests/node-postgres"
   },
-  "main": "app.js",
+  "main": "main.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
@@ -23,9 +23,9 @@
   "license": "Apache-2.0",
   "dependencies": {
     "app": "^0.1.0",
-    "pg": ">7.17.1",
-    "uuid": "^3.4.0",
+    "async-csv": "^2.1.3",
     "chai": "^4.2.0",
-    "csv-parser": "^2.3.2"
+    "pg": "^8.5.1",
+    "uuid": "^3.4.0"
   }
 }
