+++ tests/client_tests/node-postgres/run.sh
@@ -1,4 +1,4 @@
 #!/usr/bin/env bash
 
 cr8 run-crate latest-nightly \
-  -- @node app.js {node.addresses.psql.host} {node.addresses.psql.port}
+  -- @node main.js {node.addresses.psql.host} {node.addresses.psql.port}
